feat(AHWR-319): test refactor
