+++ test/unit/event/event.test.js
@@ -1,7 +1,6 @@
 const queryEntities = require('../../../ffc-ahwr-event/azure-storage/query-entities')
 const { saveEvent } = require('../../../ffc-ahwr-event/event/event')
 const mockContext = require('../../mock/mock-context')
-const { buildRowKey } = require('../../../ffc-ahwr-event/event/event-util')
 
 jest.mock('../../../ffc-ahwr-event/azure-storage/query-entities')
 
@@ -58,24 +57,6 @@ describe('Event function', () => {
     await saveEvent(mockContext, message)
     expect(queryEntities).toHaveBeenCalledTimes(1)
     expect(mockContext.bindings).toHaveProperty('tableBinding')
-  })
-
-  test('should construct the rowKey correctly', () => {
-    const partitionKey = '123456789'
-    const timespan = 1698670400000
-    const eventType = 'event'
-    const expectedRowKey = '123456789_1698670400000_event'
-    const result = buildRowKey(partitionKey, timespan, eventType)
-    expect(result).toBe(expectedRowKey)
-  })
-
-  test('should handle empty partitionKey, timespan, and eventType gracefully', () => {
-    const partitionKey = ''
-    const timespan = ''
-    const eventType = ''
-    const expectedRowKey = '__'
-    const result = buildRowKey(partitionKey, timespan, eventType)
-    expect(result).toBe(expectedRowKey)
   })
 
   test('rowKey is constructed correctly', async () => {
